New: [AEA-4563] - Create trackerUserInfo Lambda (#182)
## Summary

- :sparkles: New Feature

### Details

When a user is calling the CIS2 API with the Cognito session token CIS2
should return some details around the users profile.

It should  retrieve the role_name, org_code and activity_code:

- The org_code will be used to call DoHS (see
[AEA-4555](https://nhsd-jira.digital.nhs.uk/browse/AEA-4555)) to get
details of the organisation. (for now, just return the org_code -
fetching information from DoHS is the other ticket)
- The role_name and activity_code will give details around the role
profile name and whether that has access to the CPTS.