Fix: [AEA-0000] - Fix header behaviour (#339)
## Summary

- Routine Change

### Details

The header is becoming a bit of a mess. It's time that got sorted out,
so I've refactored a bit to make it more well-behaved. Since I was doing
some tidying up anyway, I've also fixed a few minor bugs here and there
and made sure to lint the files I've touched, so this commit looks
bigger than it is! Things like requests being made multiple times in
parallel, and redirecting causing unintended content to flash on the
screen, nothing massive.

In addition, I've also made an alternative version of `useState` that
mirrors its data to the cache. This is to protect certain data from
being purged when the page reloads. It's used like so:
```javascript
import React from 'react'
import { useLocalStorageState } from '@/helpers/useLocalStorageState'

export default function ExampleComponent() {
  const [var, setVar] = useLocalStorageState("storage_key", "storage_group", true);

  const updateVar = () => {
    if (var) {
      setVar(false)
    } else {
      setVar(true)
    }
  }

  return (<div />)
}
```
When the variable is updated, the new value is stored in the cache. When
the page is reloaded, the default value is fetched from the cache until
it is changed, or fetched from the backend.

## Header link behaviour:
```
Exit:
Shown if:
	- we are on the logout page and we are not signed in
	- we are on the select your role page, AND have no roles with access

Logout:
Shown if:
	- we are logged in

Select Your Role:
Shown if:
	- We are not on the select your role page
	- We are not on the change role page
	- We are signed in
	- We do not have a selected role already

Change Role:
Shown if:
	- We are not on the select your role page
	- We are not on the change role page
	- We are signed in
	- We DO have a selected role already
```