Chore: [AEA-0000] - Axios migration (#375)
## Summary

- Routine Change

### Details

~~Occasionally, the tracker user info request fails (I think with a 502
error, though this is captured outside of regression tests). I want to
track down the problem, and this is the sandbox I'll use to do it. There
is a counterpart regression test PR here:
https://github.com/NHSDigital/electronic-prescription-service-api-regression-tests/pull/224~~

This PR migrates us over to using Axios for our API requests. For now,
the only real benefit is that I've added an interceptor that retries
requests a few times if they fail, but in the future we may want more of
this kind of flexibility.

I've also tracked down and fixed one or two bugs - namely, the issue
with a empty selected role causing the RBAC banner and pre-selected role
elements to show up when they shouldn't, and an issue where the logout
page eagerly redirected to the login page as soon as the user was logged
out. Also, some cleaning up on the regression test side of thigs, too.
Oh, and linting.