New: [AEA-4729]  - GET /prescriptionList (#385)
## Summary

Once a user has used one of three search criteria's to find a
prescription they will then be presented with the results of that
search. This will initially be in a high level list which will give some
basic details.

Depending on the search they will either be presented with prescriptions
or won't be.

- :sparkles: New Feature

This is what the response of a successful prescription search looks
like:

![image](https://github.com/user-attachments/assets/dce259a0-4e74-4665-8ff4-60678baec7ac)

---------

Co-authored-by: Bence Gadanyi <bence.gadanyi1@nhs.net>