New: [AEA-4778] - Prescriptions list page (#572)
## Summary

- :sparkles: New Feature - added the Prescription List place-holding
page with some minimal functionality to be able to traverse back to the
previous page with the correct tabs being selected, depending on whether
we visited the page from the nhs number or prescription id search tabs.
- :warning: Potential issues that might be caused by this change - as a
part of this PR I also renamed a few of our cpts-ui pathnames, which
affects not only the frontend package but so do any authentication flows
that redirect us back to the cpts-ui project.