Fix: [AEA-5046] - Change Role page showing duplicates and issues on refresh (#527)
## Summary

Issue 1: The /changerole page is showing the selected role in the roles
with access list when using the changerole links within the UI. After
refreshing the page the duplicate is removed from the roles with access
list and shows the page as expected. 

Issue 2: The bug from
[AEA-4974](https://nhsd-jira.digital.nhs.uk/browse/AEA-4974) is also
back, but only after refreshing the /changerole page will you see the
'No Access text'. This is due to the user only having 1 role with access
which is selected, the user has an empty roles with access list.

- Routine Change