New: [AEA-4799] - Prescribed and Dispensed Items (#693)
## Summary

🎫 [AEA-4799](https://nhsd-jira.digital.nhs.uk/browse/AEA-4799)
Prescribed and Dispensed Items
🧪 Regression Tests:
https://github.com/NHSDigital/electronic-prescription-service-api-regression-tests/pull/294

:sparkles: New Feature

### Details

In the prescription details page this will give information on the items
which have been prescribed and subsequently dispensed. If there has been
a change in the medication, quantity and dosage instruction from what
has been dispensed to prescribed, it will have a drop down on what has
been prescribed.