New: [AEA-5182] - cookie banner (#800)
- :sparkles: New Feature


It is a requirement as part of UK-GDPR we must give the users the
ability to turn off 'optional' cookies. This is for things like
analytics which we track to improve User Experience.

In the NHS Design Guidelines and .Gov guidelines they do this in the
form of a banner