New: [AEA-5360] - Create the Patient not found page (#928)
## Summary

🎫 [AEA-5360](https://nhsd-jira.digital.nhs.uk/browse/AEA-5360) Create
the Patient not found page
🎫 [AEA-5361](https://nhsd-jira.digital.nhs.uk/browse/AEA-5361) Update
the 'Too many search results returned' page
🧪 Regression Tests:
https://github.com/NHSDigital/electronic-prescription-service-api-regression-tests/pull/338

:sparkles: New Feature

### Details

The basic details search does a search in PDS and depending what has
been searched it might return no patient back. With this it should in
the UI display 'Patient not found' so the user can do a search again. It
should also show this page for S-flag patients as we aren't showing them
in the Prescription Tracker.

The basic details search does a search in PDS, if there has been 11 or
more returned this page should be returned. When a user goes back should
store the original search.

This pull request implements the 'Patient not found' page and updates
'Too many search results' page.