New: [AEA-5481] - Base path redirect to /site (#1056)
## Summary

- :sparkles: New Feature

### Details

Redirect users landing on the base path of '/' to /site.
Include redirects for first segment paths, so that '/foo' becomes
'/site/foo' in an attempt to allow the user to find a potentially valid
page.
Send the user a 404 to any pages with 2 segments.
Ensure pre-defined path rules happen first, so that APIs work.

---------

Signed-off-by: Connor Avery <214469360+connoravo-nhs@users.noreply.github.com>
Co-authored-by: anthony-nhs <121869075+anthony-nhs@users.noreply.github.com>