New: [AEA-5519] - Part 1: Create session ID and set in JWT, feed to frontend via tracker user info (#1122)
## Summary

- :sparkles: New Feature

### Details

Create a session ID value that is set within the JWT for the users
current login token.
Register that session ID against the user in token table, or a session
management table if a session already exists (for further functionality
to block concurrent sessions).

Feed understanding on whether multiple sessions exist and if the current
session is considered 'concurrent'.

---------

Signed-off-by: Connor Avery <214469360+connoravo-nhs@users.noreply.github.com>