New: [AEA-5714] - CPTS UI Missing History when Cancelling a prescription that is 'With Dispenser - Active (#1304)
## Summary

After Cancelling a prescription that is 'With Dispenser - Active' the
history column is missing the title for the cancel event in CPTS UI.

The CPTS API returns the below title, which has been found to be the
cause of the missing data in the UI.
The word 'active' is not accounted for in the backend code and is
therefore missing from the endpoint.

"title": "Prescription/item was not cancelled. With dispenser active.
Marked for cancellation"



- Routine Change

---------

Co-authored-by: Jim Wild <james.wild6@nhs.net>